requirements.txt
# Core
torch>=2.0.0
transformers>=4.35.0
tokenizers>=0.14.0
accelerate>=0.24.0
deepspeed>=0.12.0

# Data
datasets>=2.15.0
sentencepiece>=0.1.99

# Evaluation & Reporting
matplotlib>=3.7.1
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Logging
wandb>=0.16.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0.1
omegaconf>=2.3.0

# Code Execution (for RLAIF)
RestrictedPython>=6.0

# API Serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0

# Cloud (Azure)
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0

# Monitoring
psutil>=5.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
